gsupplicant: Switch to different BSSID of same SSID in case of signal bss removed
Teach ConnMan to select the next BSSID when two APs (e.g. 2.4 GHz and
5 GHz) with the same SSID and security configuration are available.

Currently when one AP disappears ConnMan will retry to connect to the
old AP point even though there is another matching BSSID available.

So when wpa_supplicant sends a remove signal we find a matching BSSID
and use it even when it's not from the same AP.