vpn: Let previous connection to finish before starting new one.
If a provider is being re-used for a VPN connection in case of
re-connection when network changes happen rapidly it is imperative to
let the provider to finish the previous connection properly. If the
provider is being disconnected, let caller know that transition is in
progress. Do the same if the last connection failed but first, do a
state state transition to idle to allow the failed connection to finish
as well.

Otherwise the reference counters are never decreased as provider goes
from disconnect/failure/ready -> connect and the old interface will
be still up. Also the routes of that interface will remain in routing
table. Until the VPN is re-connected for the second time and interface
is re-used the interface and routes will remain. This happens because
the vpn_exit() is called after the VPN has been re-connected in most
cases and when there is a reference existing, contents are not cleared
and routes are not freed, keeping the interface also up.

This commit naturally makes the re-connection a bit slower as the state
transition must finish first. The issue described happens usually when
dropping out from WiFi after hopping between 2 or more WiFi networks and
using a fallback connection (mobile data, ethernet) and the VPN gets
re-connected immediatelly after the change of network is detected.

The scenario is usually:
1) When connected to WiFi A connect manually to WiFi B
2) Let VPN connect using WiFi B
3) Shutdown WiFi B after VPN has connected
4) Let to switch to other network type (e.g, mobile data, ethernet)
5) Let VPN to autoconnect using the network type used
6) Let to automatically connect to WiFi A and to establish new VPN conn.
7) There are 2 vpn adapters up with stale routes of the old adapter