vpn: Set provider state to disconnect if driver has no disconnect impl.
This commit adds setting of provider state to disconnect when the driver
has no disconnect() function implemented. This is imperative to do in a
scenario when network conditions change rapidly and VPN is being
re-connected. This ensures the state transition from ready to disconnect
in such case when VPN driver does not have the functionality implemented
to indicate state transition.