[connman] Add access control to connman-vpn dbus interface. Contributes to JB#45379
Currently access to the connman-vpn dbus interfaces (connection and
manager) are controlled only by the policy config, which sets `<policy
at_console="true">`, allowing access to all users.

This change adds internal access control to the connection dbus
interface, which is configurable using a similar plugin approach as the
existing access method implemented for connman. It also adds a
Sailfish-specific plugin that blocks access for non-privileged users
base on a policy file installed at /etc/connman/vpn-dbus-access.conf.

The exceptions are the GetProperties and GetConnections dbus methods
which are available to all users (including nemo).

This Sailfish configuration is needed for MDM, to prevent users from
being able to connect to and disconnect from VPNs when this capability
is being restricted by the MDM policy.