vpn: Run VPN plugin binaries with configurable DAC privileges
This enables to support running VPN plugin binaries with configurable
DAC privileges. Function vpn_task_setup() is registered as a custom
setup callback function into task.c and is called when the task setup
function is called prior to running the task.

When a VPN plugin registers itself (with vpn_register()) the DAC
privilege configuration for it is loaded from config with
vpn_settings_parse_vpn_plugin_config(). This configuration is stored by
vpn-settings.c and is free'd when the VPN plugin is unregistered with
vpn_unregister(). DAC privilege configuration of a plugin is passed as
custom setup user data to connman_task_create() when vpn_connect() is
called.

DAC privilege configuration is set when a VPN plugin is to be connected
(vpn_task_setup() is called). Function vpn_task_setup() sets the DAC
privileges in following order: change group, set supplementary groups
and change user id. Use of setgid(), setgroups() and setuid() needs
CAP_SETGID and CAP_SETUID capabilities.