[connman] Check preference for connected services. Contributes to JB#43191
It is unnecessary to do a service preference check for each service.
This saves some processing time and reduces mem fragmentation as the
list from preferred_tech_list_get() does not need to be free'd.

The preference of services should be checked only for connected services
when sorting service list. The services that are not connected will not
be in the preferred technologies list. It might cause some confusion in
the logs.