[connman] Fix service preference check. Contributes to JB#43191
Ignore VPN services in service preference check. VPNs are special case
depending on other services as transport. They are never in preferred
technologies lists.

Use the preferred technology type list instead of using service_list. It
was wrong to use service_list here as service_preferred_over() is called
while sorting it and will result in undesirable results. This fixes the
sorting by simply using the service types to get their position in the
list.

Service preference is based on the position of the service type in the
preferred technology list. Lower index means that service type is
preferred over the one having higher index in the list. If the service
type is not in the preferred technology list, then the one being in that
list is preferred. Function service_preferred_over() is changed to return:
 *  1 when service b is preferred over a
 * -1 when service a is preferred over b
 *  0 when preference sorting does not apply

Connected VPNs are preferred over any other service. Non-connected VPNs
should be left as last.

Also reduce log noise on  __connman_service_is_default_route().