[connman] Home must be mounted before connmand is run. Fixes JB#46139
This forces connman to wait for /home to be mounted before it is
started. It can be achieved with RequiresMountsFor, which is equal to
having Requires and After for home.mount. Also, connman needs to be
stopped when /home is unmounted. This is achieved with PartOf.

Added a systemd service additional conf file to be put under systemd
system/connman.service.d/ which makes the above possible.