provider: Implement function to set VPN service autoconnect value
Add connman_provider_set_autoconnect() for setting the VPN autoconnect
value. If a change took place save the VPN service.

It is imperative to save the VPN service when autoconnect changes since
otherwise the change is done only to run-time service settings and not
written to disk if user has canceled VPN agent dialog, for example.
Otherwise the data within connman and service file is out of sync, and
next restart of connman will be a race between each VPNs that were
attempted to connect but canceled by user.