[vpn] Use provider as agent context in OpenVPN. Contributes to JB#45903
Use data->provider in OpenVPN as context for agent. This is similar
approach as in other plugins. Using provider as context allows to cancel
the agent request and to send reply to pending request when
vpn-provider.c:connect_cb() gets a timeout or gets no reply from the VPN
agent.