Separate is_valid_pair from is_valid_range function.
The is_valid_range function does not check the range is in ascending
order. This normally doesn't matter for 2 of the functions that call it,
using it for:
        case IPTABLES_OPTION_ESP:
                return is_valid_range(params[0], ":", NULL, 0);
In this case ESP module in kernel does a check in the spi_match
function:
        r = (spi >= min && spi <= max) ^ invert;
and a non-ascending order will not match a single packet,
and for:
        case IPTABLES_OPTION_CONNTRACK:
                /* --ctexpire has an integer or integer range ':' as sep  */
                if (option_position == IPTABLES_CONNTRACK_CTEXPIRE)
                        return is_valid_range(params[0], ":", NULL, 0);
which does a very similar check in the kernel:
                if ((expires >= info->expires_min &&
                    expires <= info->expires_max) ^
                    !(info->invert_flags & XT_CONNTRACK_EXPIRES))

The function is also used for:
 * [!] --icmp-type type[/code]  (or numeric type or type/code)
and
 * [!] --mark value[/mask]      Match nfmark value with optional mask
and these are not ranges at all.

Thus, I split the function to 2, one that adds another sanity check for
the order of the 2 numerical values, and one that checks just a pair of
strings to be numerical and/or complacent to the callback function.

------------------------------------------------------------------------

This form of a call that appears in other places of the code:

is_valid_range(params[0], ":", NULL, 0)

It checked the cb function before setting the result on the fact that
the last parameter was a RANGE_CALLBACK_AND or a RANGE_CALLBACK_OR. This
meant that this call previously acted as RANGE_CALLBACK_OR, though the 0
meant RANGE_CALLBACK_AND.

Since the new function checks if the callback exists and says that it
failed if it was provided with a RANGE_CALLBACK_AND and a NULL function
pointer, the enum order is modified to account for this simple syntax of
the call to work after the change.