dhcp: Bind to interface when sending
Having two interfaces and one goes down, ConnMan previously sent a
DHCP-release only bound to srcIP (via bind), and dstIP, (via connect).
If an interface goes down, the Linux kernel will try to route the
packet any means possible. Effect was that it routed it out on the
other available interface. It might even SNAT the srcIP due to ConnMan
NATing setup.

The result was a very strange DHCP-release. A dstIP of an DHCP-server
that the incorrect interface probably could not reach. A srcIP that
seemed correct (due to SNAT rule). An included address to release
inside the DHCP-release that was not the same as the srcIP, but rather
the actual IP to release.

This fixes this via setsockopt(.., SO_BINDTODEVICE, ...) which binds
the socket to the specific interface name, and packets are not seen on
the incorrect interface anymore.