[openvpn] Use VPN Agent for Encrypted Private Key pwd. JB#47237
React also to >PASSWORD:Need 'Private Key' request on OpenVPN management
channel, when password for encrypted private key is requested. This adds
another UI password dialog and works similarly to the username/password
request. The reply is processed and stored as hidden value. The value of
the password is written to the management channel with: 'password
\"Private Key\"' format.