[openvpn] Store Encrypted Private Key pwd in memory. JB#47237
Store the Encrypted Private Key password in vpn-provider.c as hidden
value. Disallow VPN agent to save this value, but utilize the VPN agent
dialog to retrieve the password from user.

The value needs to be cleared, if the value is queried for the second
time. This is because OpenVPN does not indicate the errors in Private
Key decryption via management channel nor it shuts the VPN down in case
of error.