[openvpn] Improve Private Key decryption detection. JB#47237
Do not add VPNAgent.AuthFailure to VPN agent request when sending
request for Private Key Password. USe failed_attempts_privatekey as
counter and increase it if management channel reports it (not working,
OpenVPN does not inform about this). Increase the same counter when
retrieving the password from memory, which is a temporary solution to
get rid of invalid password stored in memory.

Also clear the failed attempts counters for both credentials and private
key password when connection has succeeded (in ov_connect_done()).