wispr: Prevent use-after-free from agent browser request
Agent requests take a reference on the service object, but this doesn't
guarantee that the wispr context is kept alive.  When we get a callback,
lookup the context from first principles and verify that the object
we've been given is still a context on the given service.

This prevents a use-after-free on the wispr context pointer if the agent
takes a long time to respond (or fails to respond resulting in a DBus
timeout) and the context is freed before that response arrives.