vpn: Restrict connection requests to one per connection
Utilize DBusPendingCall in struct connection_data to keep track of
connection requests for VPNs. There should not be more than one pending
call per VPN at a time. If there is a pending connection request return
-EALREADY to indicate the status.

If there is a pending connection request when disconnecting the VPN
connection consume the reply by changing the notify function to
remove_connection_reply() which cleans up the call as well. If pending
call exists cancel it when connecting or disconnecting.