vpn-provider: Always reply to connect request D-Bus message
It is imperative to always send a reply/message to the D-Bus method
call. Otherwise the org.freedesktop.DBus.Error.LimitsExceeded error will
be received for the connection.

In case the VPN is in failure state its state will be changed to idle
and -EINPROGRESS is reported as will be when VPN is being disconnected.
In both of these cases a reply to the method call is to be sent to
avoid leaving pending replies hanging in D-Bus.

Also, set VPN driver state to idle when a VPN is attempted to be
connected in failure state. This ensures that state is set for both
provider and the driver impl.