[openconnect] Force task to use C locale with screen scraping. JB#45678
This is a temporary fix for OpenConnect VPN plugin to force C locale
for the task because screen scraping is used to get information on the
process. This is required because when the user is changed to other
than root the translations are enabled if language settings define so.

This should be reverted/removed when OpenConnect uses the library for
authenticating. This is mainly aimed for 3.2.1 release, where
OpenConnect user is changed to "nemo".