vpn-provider: Reply to connect request only once
It is imperative to reply to the D-Bus method call only once. The
-EINPROGRESS returned by __vpn_provider_connect() should not be
immediately replied with an appropriate error. This is because the
connect_cb() will also send a reply based on the error returned by
the VPN plugin, or an empty one if there is no error.

In case the VPN is in failure state its state will be changed to idle
and -EINPROGRESS is reported as will be when VPN is being disconnected.
In both of these cases a reply to the method call is to be sent
directly.

Also, set VPN driver state to idle when a VPN is attempted to be
connected in failure state. This ensures that state is set for both
provider and the driver impl.