[connman] Add CVE-2019-14899 IPv6 mitigation. Contributes to JB#48350
Basic IPv4 mitigation is to have net.ipv4.conf.all.rp_filter = 1 and for
IPv6 following rule was suggested in [1]:
  ip6tables -t mangle -I PREROUTING -m rpfilter --invert -j DROP

This is a temporary solution until other than filter table rules are
supported in firewall configs. Adding the rule to the general config
makes it operate as other rules read from configs, to be added at
startup and removed when shutdown. Uses connman-PREROUTING table, since
is added as managed rule.

[1] https://seclists.org/oss-sec/2019/q4/123