vpn-provider: Use SetProperty to set multiple property values
To comply upstream request use SetProperty to set multiple property
values as well. This will remove SetProperties from use.

If the given D-Bus message contains DBUS_TYPE_ARRAY as variant, which is
not containing User Routes attempt to process it as a dict of
properties. Use set_properties() for processing and return the message,
or error, it creates back to sender.

Allow also to clear values with SetProperty call.