vpn-provider: Implement setting of multiple VPN properties with one call
Enhance SetProperty D-Bus method to accept an array of VPN properties in
order to enable setting of multiple properties with one D-Bus call. For
enabling this set_properties() method is implented, and is called when
an array is received that does not have name "UserRoutes". This improves
the performance when multiple (or all) properties are to be changed but
reduces the granularity of errors reported back to the caller. Properties
are accepted in the same format as with GetProperties D-Bus method sends
them, a{sv}, wrapped inside a variant value.

When an array of properties is sent each changed property value emits a
PropertyChanged signal. If there is at least one changed property
value then the VPN provider config is written to disk. Disk write is
done after all properties in the array are processed.

Empty strings as property values, or an empty array for "UserRoutes" are
allowed with this change when using SetProperty. In both cases the
effect is the same as calling ClearProperty on a property. Thus, this
change allows to clear multiple properties with one call when sending an
array of (empty) properties.

If there is at least one invalid property in the array then
InvalidProperty D-Bus error is sent as reply. Invalid properties >
immutable properties in case of error and there can be permission errors
caused by immutable properties when InvalidProperty error is sent. Only
when there are only permission errors then PermissionDenied D-Bus error
is sent. In both cases the D-Bus error message is amended with the
property names that caused the errors. If there are also properties that
had proper values and the values were changed a PropertyChange signal is
sent for each and VPN provider settings are written to disk.

Move setting of a single property into its own function, usable by both
set_property() and set_properties(). The set_vpn_property() retains the
functionality by changing either the routes or a single string value,
returns 0 when success, -EALREADY when there is no change on property
value and appropriate error (-EINVAL/-EPERM/-ENOMEM) otherwise.