service: Store configuration in empty keyfile
Instead loading the keyfile from disk with the old configuration,
start with an empty configuration. Because we do not have a previous
configuration which needs to be updated we can drop all the remove
code. This avoids complex code paths when switching from one
connnection method to the next one (dhcp to manual and back).